Modifying for the Web Browser.
I modified a tiny bit of code to make it run just as easily in the web
browser. I also removed some side-effects that maximized the size of the
videos. This may or may not have been the intended effect, but I did not
want it.

I also provided a minified version of the file. If I did something
wrong, please correct me.